+++ src/db/QueryBuilder.php
@@ -1511,21 +1511,19 @@ class QueryBuilder extends \yii\base\BaseObject
      */
     public function buildWithQueries($withs, &$params)
     {
-        if (empty($withs)) {
+        if ($withs === []) {
             return '';
         }
 
         $recursive = false;
         $result = [];
 
-        foreach ($withs as $i => $with) {
-            if ($with['recursive']) {
-                $recursive = true;
-            }
+        foreach ($withs as $with) {
+            $recursive = $with['recursive'] ?? $recursive;
 
             $query = $with['query'];
             if ($query instanceof Query) {
-                list($with['query'], $params) = $this->build($query, $params);
+                [$with['query'], $params] = $this->build($query, $params);
             }
 
             $result[] = $with['alias'] . ' AS (' . $with['query'] . ')';